+++ src/main/gen/io/runescript/plugin/lang/psi/RsLongLiteralExpression.java
@@ -1,0 +1,13 @@
+// This class is automatically generated. Do not edit.
+package io.runescript.plugin.lang.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public interface RsLongLiteralExpression extends RsExpression {
+
+  @NotNull
+  PsiElement getLong();
+
+}