+++ src/main/gen/io/runescript/plugin/lang/psi/impl/RsLongLiteralExpressionImpl.java
@@ -1,0 +1,36 @@
+// This class is automatically generated. Do not edit.
+package io.runescript.plugin.lang.psi.impl;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.lang.ASTNode;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiElementVisitor;
+import com.intellij.psi.util.PsiTreeUtil;
+import static io.runescript.plugin.lang.psi.RsElementTypes.*;
+import io.runescript.plugin.lang.psi.*;
+
+public class RsLongLiteralExpressionImpl extends RsExpressionImpl implements RsLongLiteralExpression {
+
+  public RsLongLiteralExpressionImpl(@NotNull ASTNode node) {
+    super(node);
+  }
+
+  @Override
+  public void accept(@NotNull RsVisitor visitor) {
+    visitor.visitLongLiteralExpression(this);
+  }
+
+  @Override
+  public void accept(@NotNull PsiElementVisitor visitor) {
+    if (visitor instanceof RsVisitor) accept((RsVisitor)visitor);
+    else super.accept(visitor);
+  }
+
+  @Override
+  @NotNull
+  public PsiElement getLong() {
+    return notNullChild(findChildByType(LONG));
+  }
+
+}