+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [Unreleased]
 
+## [1.5.0] - 2023-11-27
+
 ### Added
+
 - Add support for `long` numeric type.
 - Add `stylesheet` to the valid types list.
 - Add type checking for arithmetic operators.
@@ -88,7 +91,8 @@
 
 - The new file action no longer shows on top of everything else.
 
-[Unreleased]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.4.1...HEAD
+[Unreleased]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.4.1...v1.5.0
 [1.4.1]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/waleedyaseen/intellij-runescript/compare/v1.2.0...v1.3.0